test/gtest/bn_activ_infer: added separate tester class for batchnorm inferences
Added a separate tester class for batchnorm inferences so we can reuse it in other batchnorm inference tests. Also modified the existing batchnorm inference test to use this tester class.
